+++ tests/bootstrap.php
@@ -63,3 +63,4 @@ yourls_add_action( 'pre_yourls_die', function($params) {
 echo "YOURLS installed, starting PHPUnit\n\n";
 
 require_once __DIR__ . "/tests/auth/AbstractLoginTestCase.php";
+require_once __DIR__ . "/tests/auth/LoginAssertionTrait.php";
